+++ SEN381 Pr/Presentation Layer/JobsFrm.cs
@@ -21,19 +21,44 @@ namespace SEN381_Pr
 
         private void btnInsert_Click(object sender, EventArgs e)
         {
-            Con.InsertJobData(dgvJobs,txtJobID.Text,txtAddress.Text,txtReference.Text,txtLevel.Text,txtTech.Text,txtContract.Text);
+            if (txtAddress.Text == string.Empty || txtReference.Text == string.Empty || txtLevel.Text == string.Empty || txtTech.Text == string.Empty || txtContract.Text == string.Empty)
+            {
+                MessageBox.Show("Please enter all fields!");
+            }
+            else
+            {
+                Con.InsertJobData(dgvJobs, txtJobID.Text, txtAddress.Text, txtReference.Text, txtLevel.Text, txtTech.Text, txtContract.Text);
+
+            }
         }
 
         private void btnUpdate_Click(object sender, EventArgs e)
         {
+            if (txtJobID.Text == string.Empty)
+            {
+                MessageBox.Show("Please select an ID to update!");
+            }
+            else if(txtAddress.Text == string.Empty || txtReference.Text == string.Empty || txtLevel.Text == string.Empty || txtTech.Text == string.Empty || txtContract.Text == string.Empty)
+            {
+                MessageBox.Show("Please enter all fields!");
+            }
+            else
+            {                
 
+            }
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            Con.DeleteJobData(dgvJobs, txtJobID.Text, txtAddress.Text, txtReference.Text, txtLevel.Text, txtTech.Text, txtContract.Text);
+            if (txtJobID.Text == string.Empty)
+            {
+                MessageBox.Show("Please select an ID to escelate!");
+            }
+            else
+            {
+                Con.DeleteJobData(dgvJobs, txtJobID.Text, txtAddress.Text, txtReference.Text, txtLevel.Text, txtTech.Text, txtContract.Text);
+            }
         }
-
         private void JobsFrm_Load(object sender, EventArgs e)
         {
             dgvJobs.AutoGenerateColumns = true;
@@ -76,10 +101,16 @@ namespace SEN381_Pr
             }
 
         }
-
         private void btnEscelate_Click(object sender, EventArgs e)
         {
-            Con.EscelateJob(dgvJobs, txtJobID.Text, txtAddress.Text, txtReference.Text, txtLevel.Text, txtTech.Text, txtContract.Text);
+            if (txtJobID.Text == string.Empty)
+            {
+                MessageBox.Show("Please select an ID to escelate!");
+            }
+            else
+            {
+                Con.EscelateJob(dgvJobs, txtJobID.Text, txtAddress.Text, txtReference.Text, txtLevel.Text, txtTech.Text, txtContract.Text);
+            }            
         }
 
         private void dgvJobs_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -87,8 +118,8 @@ namespace SEN381_Pr
             DataGridViewRow row = dgvJobs.CurrentRow;
             if (!row.IsNewRow)
             {
-                txtJobID.Text = row.Cells["JobId"].Value.ToString(); ;
-                txtContract.Text = row.Cells["ContractId"].Value.ToString(); ;
+                txtJobID.Text = row.Cells["JobId"].Value.ToString(); 
+                txtContract.Text = row.Cells["ContractId"].Value.ToString(); 
                 txtLevel.Text = row.Cells["EscalationLevel"].Value.ToString();
                 txtReference.Text = row.Cells["ReferenceNumber"].Value.ToString();
                 txtAddress.Text = row.Cells["AdressId"].Value.ToString();
